+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the tool_vault plugin will be documented in this file.
 
 ## Unreleased
 
+### Added
+- Callback to allow custom distributions to execute code after each table is restored #17
 ### Fixed
 - Files over 5Gb were rejected by the storage provider. They are now uploaded using
   multipart upload.